+++ docs/docs.go
@@ -3087,6 +3087,12 @@ const docTemplate = `{
                         "required": true
                     },
                     {
+                        "type": "string",
+                        "description": "The name of the service which invokes the API",
+                        "name": "service",
+                        "in": "query"
+                    },
+                    {
                         "description": "body data",
                         "name": "data",
                         "in": "body",
@@ -3725,6 +3731,59 @@ const docTemplate = `{
                                 "$ref": "#/definitions/getEventUserIDsResponse"
                             }
                         }
+                    }
+                }
+            }
+        },
+        "/group/{group-id}/members": {
+            "put": {
+                "security": [
+                    {
+                        "AppUserAuth": []
+                    }
+                ],
+                "description": "Updates a membership. Only the status can be changed.",
+                "consumes": [
+                    "application/json"
+                ],
+                "produces": [
+                    "application/json"
+                ],
+                "tags": [
+                    "Admin"
+                ],
+                "operationId": "AdminCreateMemberships",
+                "parameters": [
+                    {
+                        "type": "string",
+                        "description": "APP",
+                        "name": "APP",
+                        "in": "header",
+                        "required": true
+                    },
+                    {
+                        "description": "body data",
+                        "name": "data",
+                        "in": "body",
+                        "required": true,
+                        "schema": {
+                            "type": "array",
+                            "items": {
+                                "$ref": "#/definitions/model.MembershipStatus"
+                            }
+                        }
+                    },
+                    {
+                        "type": "string",
+                        "description": "Group ID",
+                        "name": "group-id",
+                        "in": "path",
+                        "required": true
+                    }
+                ],
+                "responses": {
+                    "200": {
+                        "description": "OK"
                     }
                 }
             }
@@ -4385,6 +4444,13 @@ const docTemplate = `{
                     "description": "core user net id",
                     "type": "string"
                 },
+                "net_ids": {
+                    "description": "core user net ids",
+                    "type": "array",
+                    "items": {
+                        "type": "string"
+                    }
+                },
                 "offset": {
                     "description": "result offset",
                     "type": "integer"
@@ -4959,6 +5025,18 @@ const docTemplate = `{
                     "type": "boolean"
                 },
                 "reject_reason": {
+                    "type": "string"
+                }
+            }
+        },
+        "model.MembershipStatus": {
+            "type": "object",
+            "properties": {
+                "net_id": {
+                    "type": "string"
+                },
+                "status": {
+                    "description": "pending, member, admin, rejected",
                     "type": "string"
                 }
             }